Fixes Tasks async/await issue. Introduced to ConfigureAwait framework.
GUI used to freeze when working with the async/await methods using the TPL library.
So using the ConfigureAwait(false) sort this thread synchronization contexting problem when the await hangs while calling from the ui thread.
This framework just helps us to ConfigureAwait at the assembly level instead of writing this in every task.